#23 Make emailVerification folder that has EmailVerification component and its style file  and Use it in Email.tsx
