+++ src/emailVerification/style.scss
@@ -1,5 +1,5 @@
-@import '../../common.scss';
-#email {
+@import '../common.scss';
+.email-verification {
   .email-from,
   .authNumber-form,
   #inputForm-email {
