+++ src/signup/style.scss
@@ -1,4 +1,3 @@
-@import './assets/email';
 @import './assets/genderAndBirth';
 @import './assets/inputForm';
 @import './assets/pw';
